backend1/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-pymongo==2.3.0
python-decouple==3.8
tensorflow==2.18.0
joblib==1.4.2
pymongo==4.12.0
werkzeug==3.0.4